chore: move getTeamMemberEmailFromCrm to package/lib
