+++ apps/web/lib/__tests__/getTeamMemberEmailFromCrm.test.ts
@@ -7,9 +7,8 @@ import { getCRMContactOwnerForRRLeadSkip } from "@calcom/app-store/_utils/CRMRou
 import bookingFormHandlers from "@calcom/app-store/routing-forms/appBookingFormHandler";
 import { ROUTING_FORM_RESPONSE_ID_QUERY_STRING } from "@calcom/app-store/routing-forms/lib/constants";
 import { RouteActionType } from "@calcom/app-store/routing-forms/zod";
+import { getTeamMemberEmailForResponseOrContactUsingUrlQuery } from "@calcom/lib/server/getTeamMemberEmailFromCrm";
 import { SchedulingType } from "@calcom/prisma/enums";
-
-import { getTeamMemberEmailForResponseOrContactUsingUrlQuery } from "../getTeamMemberEmailFromCrm";
 
 vi.mock("@calcom/app-store/routing-forms/appBookingFormHandler", () => ({
   default: {
