+++ apps/web/lib/team/[slug]/[type]/getServerSideProps.tsx
@@ -168,7 +168,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
   const fromRedirectOfNonOrgLink = context.query.orgRedirection === "true";
   const isUnpublished = team.parent ? !team.parent.slug : !team.slug;
   const { getTeamMemberEmailForResponseOrContactUsingUrlQuery } = await import(
-    "@calcom/web/lib/getTeamMemberEmailFromCrm"
+    "@calcom/lib/server/getTeamMemberEmailFromCrm"
   );
   const {
     email: teamMemberEmail,
