+++ packages/lib/server/index.ts
@@ -9,3 +9,4 @@ export { getTranslation } from "./i18n";
 export { getDefaultLocations } from "./getDefaultLocations";
 export { default as perfObserver } from "./perfObserver";
 export { getRoutedUrl } from "./getRoutedUrl";
+export { getTeamMemberEmailForResponseOrContactUsingUrlQuery } from "./getTeamMemberEmailFromCrm";
