+++ packages/platform/libraries/index.ts
@@ -32,6 +32,7 @@ import getAllUserBookings from "@calcom/lib/bookings/getAllUserBookings";
 import { symmetricEncrypt, symmetricDecrypt } from "@calcom/lib/crypto";
 import getBulkEventTypes from "@calcom/lib/event-types/getBulkEventTypes";
 import { getRoutedUrl } from "@calcom/lib/server/getRoutedUrl";
+import { getTeamMemberEmailForResponseOrContactUsingUrlQuery } from "@calcom/lib/server/getTeamMemberEmailFromCrm";
 import { getTranslation } from "@calcom/lib/server/i18n";
 import { MembershipRole } from "@calcom/prisma/enums";
 import { credentialForCalendarServiceSelect } from "@calcom/prisma/selects/credential";
@@ -236,3 +237,5 @@ export { getBulkEventTypes };
 export { getBookingFieldsWithSystemFields };
 
 export { getRoutedUrl };
+
+export { getTeamMemberEmailForResponseOrContactUsingUrlQuery };
